chore(cz): get commitizen to work for changelog
making it work the way the lerna setup
worked before with cz-lerna-changelog now in nx repo
