+++ .husky/pre-commit
@@ -1,0 +1,1 @@
+# Lint-staged is now handled in prepare-commit-msg
