+++ .husky/prepare-commit-msg
@@ -1,0 +1,22 @@
+#!/bin/sh
+# This hook runs before the commit message is created
+# We'll generate our own commit message and skip Git's editor
+
+# Check if we already have a custom commit message (avoid infinite loop)
+if [ -f .git/COMMIT_EDITMSG ] && grep -q "Affected packages:" .git/COMMIT_EDITMSG; then
+  # We already have our custom message, use it
+  exit 0
+fi
+
+# Run lint-staged first
+echo "Running lint-staged..."
+pnpm exec lint-staged
+
+# If lint-staged passed, run our custom commit script
+if [ $? -eq 0 ]; then
+  echo "Lint-staged passed! Generating commit message..."
+  exec < /dev/tty && pnpm run commit "$1" || true
+else
+  echo "Lint-staged failed. Commit aborted."
+  exit 1
+fi
