+++ package.json
@@ -34,24 +34,19 @@
     "chore:dependencies:root": "npm update --dev --save",
     "chore:dependencies": "npm run chore:clean && npm run chore:dependencies:root && npm run chore:dependencies:packages",
     "chore:changelog": "conventional-changelog -p angular -i CHANGELOG.md -s",
-    "commit": "git-cz",
-    "version": "cz bump"
+    "commit": "node scripts/commit.js",
+    "version": "cz bump",
+    "prepare": "husky"
   },
+  "type": "module",
   "config": {
     "commitizen": {
-      "path": "cz-customizable"
-    }
-  },
-  "husky": {
-    "hooks": {
-      "prepare-commit-msg": "exec < /dev/tty && git cz --hook || true",
-      "pre-commit": "lint-staged"
+      "path": "./scripts/commit.js"
     }
   },
   "lint-staged": {
     "*.{js,json,css,md}": [
       "prettier --write",
-      "git add",
       "jest --bail --findRelatedTests"
     ]
   },
@@ -155,7 +150,6 @@
     "commitizen": "^3.1.2",
     "conventional-changelog": "^3.1.25",
     "conventional-changelog-cli": "^2.2.2",
-    "cz-customizable": "^7.4.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
     "enzyme-to-json": "^3.6.2",
@@ -164,12 +158,13 @@
     "eslint-plugin-babel": "^5.3.1",
     "eslint-plugin-prettier": "^5.5.3",
     "gitbook-updated-cli": "^2.3.4",
-    "husky": "^0.15.0-rc.13",
+    "husky": "^9.1.7",
+    "inquirer": "12.9.4",
     "jest": "^27.5.1",
     "jest-environment-jsdom": "^27.5.1",
     "jest-enzyme": "^7.1.2",
     "lerna-update-wizard": "^0.9.4",
-    "lint-staged": "^7.3.0",
+    "lint-staged": "^16.1.5",
     "nodemon": "^1.19.4",
     "nx": "21.3.7",
     "plop": "^2.7.6",
